Fix statusBarStyle handling for center window
Cause the center window never gets added to the `containedWindows`
property on TiRootViewController, statusbar style changes on center
window are ignored. More specifically the left/rights drawers window is
asked for the statusbar style by TiRootViewController.
This is fixed by calling `windowWillOpen` and `windowDidOpen` on the
center window proxy AFTER fetching the view controllers for the left and
right drawer. Cause fetching the left and right drawer view controllers
is implicitly calling `windowWillOpen` and `windowDidOpen` on their
proxies, adding them to the TiRootViewControllers `containedWindows`
array.